cuidador y mascota service

+++ src/main/java/com/dbp/pet_journey/cuidador/infraestructure/CuidadorRepository.java
@@ -3,5 +3,5 @@ package com.dbp.pet_journey.cuidador.infraestructure;
 import com.dbp.pet_journey.cuidador.domain.Cuidador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CuidadorRepository extends JpaRepository<Cuidador, Integer> {
+public interface CuidadorRepository extends JpaRepository<Cuidador, Long> {
 }
